Create a more structured search frontier
This fixes what I think is a bug; it was possible for us to explore
patterns of weight W+n before all patterns of weight W. This means there
could be primes of length W+1 that we haven't seen yet.

Also it'll let me iterate up to a max weight, instead of an iteration
count.