Add Sequence to the first stage of the search
This introduced some panics when converting simple sequences that were
too large. So I also had to modify from_family to try_from_family, and
handle the case where we can't convert. We can delay judgment in the
first stage, but in the second, we have no choice but to be mad about
it.