Add another type of composite test
This also doesn't help a huge amount... maybe it's not worth it. It's
down from 10881 branches to 10879 branches, a decrease of only 2 :(

That said, I realized I should add tests to the composite check! That'll
make it way easier to figure out if there's bugs in them. We have
examples from Curtis Bright's paper, after all.