Introduce some bit-mixing
One recurring problem we have is that sometimes we will end up splitting
the same family over and over again when a different approach would work
better. One way (though not the only way) this occurs, is when we keep
exploring on the same core repeatedly.

This can be fixed by picking a random number each time we want to do
something, but I would like this to be deterministic. So instead, we can
do some bit mixing in order to jumble up our counter and get a less
regular sequence that we can use for "random" decisions.

It would be nice to maybe use this for splitting as well!