Introduce BigSequence
Okay, figured out the problem! There are some branches that are fairly
long (in base 19, there's MMMMMMMMMMMMML*) that don't fit into a u64
when converted to a sequence. Previously, we were searching to pretty
long lengths, and these got eliminated via containing a prime. But the
recent change to split ordering means we finish the first stage faster,
and so we don't get a chance to eliminate these in that stage!

To fix this, we can change the composite checks for simple families to
use big ints, so there's no limit on the size of simple families we can
check.

This doesn't give a good way to deal with these families in the
intermediate/second stages, but we'll get there in a moment.

For now, this at least fixes the tests for base 19. Bases 23 and 28 are
still broken.