Cache small prime buffer
Well, we're not getting any closer to solving base 29, so let's work on
profiling things again.

A fair amount of our time in `test_for_prime` is spent in
`NaiveBuffer::new()`, so let's just make one and keep re-using it.

Running the first stage only for base 17, we make 44166 primality tests.
Before, it takes about 110ms, and now it takes 75ms.