Reduce the number of substring tests
If we have a lot of primes and a lot of branches, that's a lot of
substring checks we need to make. If we track which primes _could_ be
found in any particular family, that can cut down the number of checks
we need to make.

Example: we make 15x fewer checks in base 23! (Simple families don't
fare as well, but it's still an improvement.)
```
cargo run --release search 23 --first-stage-only
```

Before:
```
783865950 substring tests (6189ms)
270012017 simple substring tests (4714ms)
```

After:
```
52075479 substring tests (464ms)
75504529 simple substring tests (1582ms)
```