Check for forbidden sandwiches
There's some long derivation chains we're having trouble getting rid
of. For example, in base 29, we have trouble removing
`P000...000[0R]*000...000C`. We expand on that core, and the branch
where we produce an `R` is easily killed. But the branch where we
produce a `0` just puts us back where we started.

Since `PR0R` is prime, we can take advantage of that to split this
family into:
- families with no Rs (i.e., `P0*C`, which is simple)
- families with one R (`P0*R0*C`, which is even)
- families with more Rs (`P0*RR*R0*C`, which ???)

Not sure this is better. It means our final weight for base 29 is 54,
instead of 118, but it's still 1.3M branches total, and we didn't make a
dent in the total time.